Fix edge case where interface can get stuck unjoined (#5)
There is apparently an edge case where the changed and broken handlers
can get queued in a hook and fire in that order leading to no subsequent
changed handlers being queued and the flags all being left cleared. It's
safer overall to include joined in the handler conditions so that even
if something strange happens, it will sort itself out into the stable
state.